+++ activerecord/test/models/car.rb
@@ -9,7 +9,7 @@ class Car < ActiveRecord::Base
 
   has_many :tyres
   has_many :engines, :dependent => :destroy, inverse_of: :my_car
-  has_many :wheels, :as => :wheelable, :dependent => :destroy
+  # has_many :wheels, :as => :wheelable, :dependent => :destroy
 
   scope :incl_tyres, -> { includes(:tyres) }
   scope :incl_engines, -> { includes(:engines) }
@@ -23,4 +23,13 @@ end
 
 class FastCar < Car
   default_scope { order('name desc') }
+end
+
+class UndestroyableCar < Car
+  before_destroy :cannot_destroy
+
+  def cannot_destroy
+    errors.add(:base, 'I cannot be destroyed')
+    throw(:abort)
+  end
 end